+++ src/commands/lint/fix.js
@@ -2,10 +2,10 @@ import sourceFiles from '../../source-files';
 import executor from '../../executor';
 import { prettierOptions } from '../../config';
 
-export default (file) => {
-  const thingsToLint = sourceFiles(file);
+export default (file, { _: unknownArgs }) => {
+  const thingsToLint = sourceFiles([file].concat(unknownArgs));
 
   executor()
-    .command('prettier-eslint', [...prettierOptions, '--write', thingsToLint])
+    .command('prettier-eslint', [...prettierOptions, '--write', ...thingsToLint])
     .run();
 };