+++ src/commands/lint/index.js
@@ -2,12 +2,10 @@ import sourceFiles from '../../source-files';
 import executor from '../../executor';
 import { eslintOptions } from '../../config';
 
-export default (file) => {
-  const thingsToLint = sourceFiles(file);
-
-  console.log('linting', thingsToLint);
+export default (file, { _: unknownArgs }) => {
+  const thingsToLint = sourceFiles([file].concat(unknownArgs));
 
   executor()
-    .command('eslint', [...eslintOptions, thingsToLint])
+    .command('eslint', [...eslintOptions, ...thingsToLint])
     .run();
 };