+++ src/source-files.js
@@ -1,13 +1,13 @@
 const isValidGlob = require('is-valid-glob');
 const readPkgUp = require('read-pkg-up');
 
-export default (file) => {
-  let thingsToLint = file;
+export default (files) => {
+  let thingsToLint = (Array.isArray(files) ? files : [files]).filter(Boolean);
 
-  if (!file) {
-    thingsToLint = process.env.npm_package_config_source;
-
-    if (!thingsToLint) {
+  if (!thingsToLint.length) {
+    if (process.env.npm_package_config_source) {
+      thingsToLint = [process.env.npm_package_config_source];
+    } else {
       // Try to find package.json nearest to where the command is being executed
       const { pkg } = readPkgUp.sync();
 
@@ -17,13 +17,20 @@ export default (file) => {
         process.exit(1);
       }
 
-      thingsToLint = pkg.config.source;
+      thingsToLint = [pkg.config.source];
     }
   }
 
-  if (!isValidGlob(thingsToLint)) {
+  const validationErrors = thingsToLint.reduce((errors, thingToLint) => {
+    if (!isValidGlob(thingToLint)) {
+      errors.push(`${thingToLint} isn't a valid file or glob pattern`);
+    }
+    return errors;
+  }, []);
+
+  if (validationErrors.length) {
     // eslint-disable-next-line no-console
-    console.error(`${thingsToLint} isn't a valid file or glob pattern`);
+    console.error(validationErrors.join('\n'));
     process.exit(1);
   }